requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
duckdb==1.3.2
greenlet==3.2.4
joblib==1.5.2
numpy==2.2.6
pandas==2.3.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
